moving docs into a separate folder from client code
