+++ src/Umbraco.Web.UI.Client/package.json
@@ -4,8 +4,7 @@
     "test": "karma start test/config/karma.conf.js --singlerun",
     "build": "gulp build",
     "dev": "gulp dev",
-    "fastdev": "gulp fastdev",
-    "docs": "gulp docs"
+    "fastdev": "gulp fastdev"
   },
   "dependencies": {
     "ace-builds": "1.4.2",
@@ -55,13 +54,10 @@
     "gulp-clean-css": "4.0.0",
     "gulp-cli": "^2.0.1",
     "gulp-concat": "2.6.1",
-    "gulp-connect": "5.6.1",
     "gulp-eslint": "^5.0.0",
     "gulp-imagemin": "^4.1.0",
     "gulp-less": "4.0.1",
-    "gulp-ngdocs": "0.3.0",
     "gulp-notify": "^3.0.0",
-    "gulp-open": "3.0.1",
     "gulp-postcss": "8.0.0",
     "gulp-rename": "1.4.0",
     "gulp-sort": "2.0.0",
