+++ src/Umbraco.Web.UI.Docs/gulpfile.js
@@ -1,6 +1,7 @@
 'use strict';
 
-var config = require('../config');
+
+var config = require('../Umbraco.Web.UI.Client/gulp/config');
 var gulp = require('gulp');
 
 var connect = require('gulp-connect');
@@ -16,28 +17,28 @@ gulp.task('docs', [], function (cb) {
         html5Mode: false,
         startPage: '/api',
         title: "Umbraco Backoffice UI API Documentation",
-        dest: 'docs/api',
-        styles: ['docs/umb-docs.css'],
+        dest: './api',
+        styles: ['./umb-docs.css'],
         image: "https://our.umbraco.com/assets/images/logo.svg"
     }
 
     return gulpDocs.sections({
         api: {
-            glob: ['src/common/**/*.js', 'docs/src/api/**/*.ngdoc'],
+            glob: ['../Umbraco.Web.UI.Client/src/common/**/*.js', './src/api/**/*.ngdoc'],
             api: true,
-            title: 'API Documentation'
+            title: 'UI API Documentation'
         }
     })
     .pipe(gulpDocs.process(options))
-    .pipe(gulp.dest('docs/api'));
+    .pipe(gulp.dest('./api'));
     cb();
 });
 
 gulp.task('connect:docs', function (cb) {
     connect.server({
-        root: 'docs/api',
+        root: './api',
         livereload: true,
-        fallback: 'docs/api/index.html',
+        fallback: './api/index.html',
         port: 8880
     });
     cb();
